Extract shared ComponentStateHooks test helpers
Three tests in this package each carried their own Map-backed
ComponentStateHooks fake. Move them to src/test/component-state-hooks.ts as
makeStateHooks (non-reactive) and makeReactiveStateHooks (reactive); both
variants stay because VirtualList uses the non-reactive one by default and opts
into the reactive one only for the tests that need a set to re-render.

The shared reactive helper is VirtualList's version, which keeps the
useSetValue/useRemoveValue action references stable — production zustand actions
are stable, and an unstable setter churns useCallback and effect deps on every
render. The LightboxCarousel copy returned a fresh closure per render, so it
moves to the higher-fidelity behavior; its tests still fail against the
unpatched component, so the regression coverage is unchanged.

The shared makeStateHooks also honors the defaultValue argument, which
ExpandablePanel's copy ignored. That matches useProperty's contract and leaves
its tests passing.

VirtualList keeps two bespoke inline fakes whose backing Map the tests read
directly; only their type annotations change.
